Font context v1: used in text atom
