+++ remotion/compositions/templates/event/EventTitle.tsx
@@ -3,8 +3,7 @@ import {Title} from '../../../design/atoms/Title';
 
 export const EventTitle: React.FC<{
 	title: string;
-	style?: React.CSSProperties;
-}> = ({title, style}) => {
+}> = ({title}) => {
 	const frame = useCurrentFrame();
 	const {fps, width} = useVideoConfig();
 
@@ -25,7 +24,6 @@ export const EventTitle: React.FC<{
 				fontSize: 100,
 				width,
 				textAlign: 'center',
-				...style,
 			}}
 		>
 			{title}
