+++ remotion/design/atoms/Text.tsx
@@ -1,13 +1,16 @@
-import React, {ReactNode} from 'react';
+import React, {ReactNode, useContext} from 'react';
+import {FontContext} from '../../../src/context/fonts/FontCountext';
 
 export const Text: React.FC<{
 	children: ReactNode;
 	style?: React.CSSProperties;
 }> = ({children, style}) => {
+	const {selectedFont} = useContext(FontContext);
+
 	return (
 		<span
 			style={{
-				fontFamily: 'Helvetica',
+				fontFamily: selectedFont,
 				height: 'max-content',
 				margin: 0,
 				padding: '0 20px',
