+++ src/context/fonts/FontProvider.tsx
@@ -2,6 +2,8 @@
 import {ReactNode, useEffect, useState} from 'react';
 import {FontContext} from './FontCountext';
 import {loadFont} from '@remotion/google-fonts/NotoSans';
+import {top250} from '../../data/fonts';
+import {loadGoogleFont} from '../../app/utils/loadFont';
 
 interface FontProviderProps {
 	children: ReactNode;
@@ -12,11 +14,13 @@ export const FontProvider: React.FC<FontProviderProps> = ({children}) => {
 		'Noto Sans'
 	);
 	loadFont();
+	const fontList = top250;
 
 	useEffect(() => {
 		const storedFont = sessionStorage.getItem('selectedFont');
 		if (storedFont) {
 			setSelectedFont(storedFont);
+			loadGoogleFont(storedFont);
 		}
 	}, []);
